feat(thread-history): Fix CMDK toLowerCase bug.
CMDK was causing the current value in our comand list to be lowercased.
This in turn passed a lowercase threadId to TRPC and broke the route.
There is now a workaround built in. This has been fixed in CMDK V1.
However CMDK was just updated and Shad has not been properly updated yet.
To maintain the integrity of our frontend for now this bandaid fix should work.