feat(vector-functionality): Add scrollbar styling
Add new package to style scrollbars using tailwind. We can get rid of this in the future, however for showing off our demo I am a fan.
The scrollbar on firefox looks pretty nice. However on Chrome it looks terrible lol.
May have to create custom scrollbars in the future.